requirements.txt
beautifulsoup4>=4.14.2
docling>=2.61.2
feedparser>=6.0.12
markdown>=3.7.0
markdownify>=0.11.6
psycopg2-binary>=2.9.11
pydantic>=2.0.0
python-dotenv>=1.2.1
requests>=2.32.5
sqlalchemy>=2.0.44
youtube-transcript-api>=1.2.3
cerebras-cloud-sdk>=1.0.0
langchain>=0.3.0
langchain-cerebras>=0.1.0
langchain-core>=0.3.0
fastapi>=0.115.0
uvicorn>=0.30.0
apscheduler>=3.10.0
